+++ fabric/CHANGELOG.md
@@ -7,6 +7,7 @@
 - Made BlockPropertyWrapperTemplates#registerBlockWithItem actually register an associated BlockItem.
 - Fixed all ModelUtil methods that use multiple textures and models (such as ModelUtil#doublePlant) attempting to look for invalid textures/models.
 - Minor cleanup in event blueprint classes.
+- (Not really a bug fix) bumped log level for `PropertyWrapper#getMappedPropertyWrappers` (+ the container field variant) down to `trace` from `debug`.
 
 ## New Features