+++ fabric/CHANGELOG.md
@@ -9,6 +9,7 @@
 - Minor cleanup in event blueprint classes.
 - (Not really a bug fix) bumped log level for `PropertyWrapper#getMappedPropertyWrappers` (+ the container field variant) down to `trace` from `debug`.
 - Fixed defaultable vegetation not working properly in general.
+- (Also not really a bug) made `StandardTagProvider` force missing tags into existence by default instead of simply removing them if both provider validation and entry-specific validation are disabled, so as to conform to deterministic behaviour.
 
 ## New Features